Delete DBSetting
